# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Enterprise Real Estate ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lead-sources:
    get:
      operationId: listLeadSources
      tags: [leadSources]
      summary: List lead-sources
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: pageSize, in: query, required: false, schema: { type: integer } }
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: companyId, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: LeadSources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadSourceListResponse"
    post:
      operationId: createLeadSource
      tags: [leadSources]
      summary: Create a LeadSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadSourceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadSource"
  /lead-sources/{id}:
    get:
      operationId: getLeadSource
      tags: [leadSources]
      summary: Get a LeadSource
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: LeadSource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadSource"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateLeadSource
      tags: [leadSources]
      summary: Update a LeadSource
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadSourceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadSource"
    delete:
      operationId: deleteLeadSource
      tags: [leadSources]
      summary: Soft-delete a LeadSource
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: pageSize, in: query, required: false, schema: { type: integer } }
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: companyId, in: query, required: false, schema: { type: string } }
        - { name: branchId, in: query, required: false, schema: { type: string } }
        - { name: sourceId, in: query, required: false, schema: { type: string } }
        - { name: assignedToUserId, in: query, required: false, schema: { type: string } }
        - { name: status, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: Leads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadListResponse"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a Lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a Lead
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a Lead
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Soft-delete a Lead
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /lead-activities:
    get:
      operationId: listLeadActivities
      tags: [leadActivities]
      summary: List lead-activities
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: pageSize, in: query, required: false, schema: { type: integer } }
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: companyId, in: query, required: false, schema: { type: string } }
        - { name: leadId, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: LeadActivities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadActivityListResponse"
    post:
      operationId: createLeadActivity
      tags: [leadActivities]
      summary: Create a LeadActivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadActivityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadActivity"
  /lead-activities/{id}:
    get:
      operationId: getLeadActivity
      tags: [leadActivities]
      summary: Get a LeadActivity
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: LeadActivity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadActivity"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateLeadActivity
      tags: [leadActivities]
      summary: Update a LeadActivity
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadActivityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadActivity"
    delete:
      operationId: deleteLeadActivity
      tags: [leadActivities]
      summary: Soft-delete a LeadActivity
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /lead-follow-ups:
    get:
      operationId: listLeadFollowUps
      tags: [leadFollowUps]
      summary: List lead-follow-ups
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer } }
        - { name: pageSize, in: query, required: false, schema: { type: integer } }
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: companyId, in: query, required: false, schema: { type: string } }
        - { name: leadId, in: query, required: false, schema: { type: string } }
        - { name: status, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: LeadFollowUps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadFollowUpListResponse"
    post:
      operationId: createLeadFollowUp
      tags: [leadFollowUps]
      summary: Create a LeadFollowUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadFollowUpInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadFollowUp"
  /lead-follow-ups/{id}:
    get:
      operationId: getLeadFollowUp
      tags: [leadFollowUps]
      summary: Get a LeadFollowUp
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: LeadFollowUp
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadFollowUp"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateLeadFollowUp
      tags: [leadFollowUps]
      summary: Update a LeadFollowUp
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadFollowUpUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadFollowUp"
    delete:
      operationId: deleteLeadFollowUp
      tags: [leadFollowUps]
      summary: Soft-delete a LeadFollowUp
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    OkResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
...
    LeadSource:
      type: object
      required: [id, companyId, code, name, nameAr, isActive, createdAt]
      properties:
        id: { type: string }
        companyId: { type: string }
        code: { type: string }
        name: { type: string }
        nameAr: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string }
    LeadSourceInput:
      type: object
      required: [companyId, code, name, nameAr]
      properties:
        companyId: { type: string }
        code: { type: string }
        name: { type: string }
        nameAr: { type: string }
    LeadSourceUpdate:
      type: object
      properties:
        companyId: { type: string }
        code: { type: string }
        name: { type: string }
        nameAr: { type: string }
    LeadSourceListResponse:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/LeadSource" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...
    LeadInput:
      type: object
      required: [companyId, code, fullName]
      properties:
        companyId: { type: string }
        branchId: { type: string }
        code: { type: string }
        fullName: { type: string }
        phone: { type: string }
        email: { type: string }
        sourceId: { type: string }
        assignedToUserId: { type: string }
        status: { type: string }
        budget: { type: string }
        notes: { type: string }
    LeadUpdate:
      type: object
      properties:
        companyId: { type: string }
        branchId: { type: string }
        code: { type: string }
        fullName: { type: string }
        phone: { type: string }
        email: { type: string }
        sourceId: { type: string }
        assignedToUserId: { type: string }
        status: { type: string }
        budget: { type: string }
        notes: { type: string }
    LeadListResponse:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Lead" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...